DAL implementations almost done

+++ HaleyAbstractionsCore/WorkFlowEngine/Interfaces/DAL/ILifeCycleDAL.cs
@@ -6,11 +6,8 @@ using Haley.Models;
 
 namespace Haley.Abstractions {
     public interface ILifeCycleDAL {
-        Task<long> InsertAsync(long instanceId, long fromStateId, long toStateId, long eventId, DbExecutionLoad load = default);
-
+        Task<long> InsertAsync(long instanceId, int fromStateId, int toStateId, int eventId, DbExecutionLoad load = default);
         Task<DbRow?> GetLastByInstanceAsync(long instanceId, DbExecutionLoad load = default);
         Task<DbRows> ListByInstancePagedAsync(long instanceId, int skip, int take, DbExecutionLoad load = default);
     }
-
-    
 }
